fix(bathymetry_store): print the child level as a number, trim test fixtures (#331)
Found running the real staged Shoals layer: the per-level progress line printed
`level  -> 7` because child_level is a uint8_t and streams as a character.

Also brings test_depth_overview back under a sane runtime. The mixed-level
fixtures were writing 16-20 real 960x960 two-band GeoTIFFs where 4-6 exercise
exactly the same assertions, and the suite had grown past the ament 60s default.
Fixtures trimmed to the minimum each assertion needs, and the target given an
explicit 300s TIMEOUT for headroom on slower machines — these are deliberately
disk-level integration tests, so the runtime is inherent, not accidental.

Part of #331.

+++ marine_bathymetry_store/src/overview_pyramid.cpp
@@ -500,7 +500,9 @@ DepthOverviewBuildResult buildDepthOverviewPyramid(
         buildLevel(children, staging, child_level, native, derived);
       const std::size_t native_here = native.countAt(static_cast<uint8_t>(level));
       if (progress != nullptr) {
-        *progress << "level " << child_level << " -> " << level << ": " <<
+        // child_level is uint8_t: without the cast it streams as a character.
+        *progress << "level " << static_cast<unsigned>(child_level) << " -> " <<
+          level << ": " <<
           counts.in << " tile(s) in, " << counts.out << " overview tile(s) out, " <<
           counts.suppressed_by_native << " left to native, " << native_here <<
           " native tile(s) already at level " << level << "\n";
